WIP: refactor CLI to use mow.cli library
