+++ distribution.go
@@ -92,7 +92,6 @@ type ByDate DistributionList
 func (a ByDate) Len() int      { return len(a) }
 func (a ByDate) Swap(i, j int) { a[i], a[j] = a[j], a[i] }
 func (a ByDate) Less(i, j int) bool {
-
 	return a[i].Label < a[j].Label
 }
