+++ go.mod
@@ -1,6 +1,7 @@
 module github.com/aalpern/luminosity
 
 require (
+	github.com/jawher/mow.cli v1.1.0
 	github.com/mattn/go-sqlite3 v1.10.0
 	gopkg.in/guregu/null.v3 v3.4.0
 )
